+++ src/NetStandard/E.Standard.Api.App/Bridge.cs
@@ -1741,6 +1741,11 @@ public class Bridge : IBridge
             _localizationNamespace = typeof(T).GetLocalizationNamespace();
         }
 
+        public ILocalizer<TClass> CreateFor<TClass>()
+        {
+            return new Localizer<TClass>(_stringLocalizer);
+        }
+
         public string Localize(string key)
         {
             var val = _stringLocalizer[$"{_localizationNamespace}.{key}"];