+++ src/NetStandard/E.Standard.Localization/Abstractions/ILocalizer.cs
@@ -2,7 +2,10 @@
 public interface ILocalizer
 {
     string Localize(string key);
+
+    ILocalizer<TClass> CreateFor<TClass>();
 }
 public interface ILocalizer<T> : ILocalizer
 {
+      
 }