+++ src/NetStandard/E.Standard.WebGIS.Tools/Identify/Extensions/ApiToolEventArgumentsExtensions.cs
@@ -2,7 +2,11 @@
 using E.Standard.WebMapping.Core;
 using E.Standard.WebMapping.Core.Api;
 using E.Standard.WebMapping.Core.Geometry;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.StaticAssets;
+using Microsoft.Identity.Client;
 using System;
+using System.Linq;
 using System.Threading;
 
 namespace E.Standard.WebGIS.Tools.Identify.Extensions;
@@ -80,4 +84,11 @@ static internal class ApiToolEventArgumentsExtensions
 
         return queryShape;
     }
+
+    static public bool UseAllIdentifyTools(this ApiToolEventArguments e, bool isMultiQuery, string[] identifyOptions)
+        => e.AsDefaultTool == true
+        && isMultiQuery == true
+        && !String.IsNullOrWhiteSpace(e["identify-map-tools"])
+        && identifyOptions.Contains("all-identify-tools");
+    
 }