+++ src/NetStandard/E.Standard.WebGIS.Tools/Identify/IdentifyServiceDesktop.cs
@@ -1,4 +1,5 @@
-﻿using E.Standard.Extensions.Compare;
+﻿using E.Standard.DependencyInjection;
+using E.Standard.Extensions.Compare;
 using E.Standard.Localization.Abstractions;
 using E.Standard.WebGIS.Tools.Extensions;
 using E.Standard.WebGIS.Tools.Identify.Abstractions;
@@ -134,18 +135,22 @@ internal class IdentifyServiceDesktop
             e.ClearMenuItemValue();
 
             IApiTool button = (IApiTool)bridge.TryGetFriendApiButton(tool, toolId);
-            if (button is IApiServerTool && !button.GetType().Equals(tool.GetType())) // dont call yourself
+            if (button?.GetType().Equals(tool.GetType()) == true)
             {
-                return ((IApiServerTool)button).OnEvent(bridge, e);
-            }
-            else if (button is IApiServerToolAsync && !button.GetType().Equals(tool.GetType())) // dont call yourself
-            {
-                return await ((IApiServerToolAsync)button).OnEvent(bridge, e);
-            }
-            else
-            {
+                // do not call yourself
                 return null;
             }
+
+            var dependencyProvider = new ToolDependencyProvider(bridge, e, localizer);
+
+            return button switch
+            {
+                { } when button.GetType().IsApiServerToolNonAsync()
+                    => Invoker.Invoke<ApiEventResponse>(button, "OnEvent", dependencyProvider),
+                { } when button.GetType().IsApiServerToolAsync()
+                    => await Invoker.InvokeAsync<ApiEventResponse>(button, "OnEvent", dependencyProvider),
+                _ => null
+            };
         }
         else
         {
@@ -162,7 +167,9 @@ internal class IdentifyServiceDesktop
 
         #endregion
 
-        if (queries == null || queries.Count() == 0)
+        if (queries == null
+            || (queries.Count() == 0 && !e.UseAllIdentifyTools(isMultiQuery, identifyOptions))
+           )
         {
             if (appendToUI)
             {
@@ -308,10 +315,7 @@ internal class IdentifyServiceDesktop
 
         #endregion
 
-        if (e.AsDefaultTool == true &&
-            isMultiQuery == true &&
-            !String.IsNullOrWhiteSpace(e["identify-map-tools"]) &&
-            identifyOptions.Contains("all-identify-tools"))
+        if (e.UseAllIdentifyTools(isMultiQuery, identifyOptions))
         {
             #region Query all other "IdentiyTools" also
 
@@ -321,7 +325,8 @@ internal class IdentifyServiceDesktop
             foreach (var toolId in e["identify-map-tools"].Split(','))
             {
                 var button = bridge.TryGetFriendApiButton(tool, toolId);
-                if (button is IIdentifyTool && (button is IApiServerTool || button is IApiServerToolAsync))
+                if (button is IIdentifyTool
+                    && button.GetType().IsApiServerTool())
                 {
                     var canIdentifyResults = await ((IIdentifyTool)button).CanIdentifyAsync(bridge, new Point(click.Longitude, click.Latitude), eventScale, availableServiceIds, availableQueryIds);
                     if (canIdentifyResults != null)